backend/requirements.txt
Flask==2.3.3
influxdb==5.3.1
python-snappy==0.7.3
gunicorn==21.2.0
prometheus-client==0.17.1
protobuf==4.24.4
python-dateutil==2.8.2
pytest==7.4.2
pytest-flask==1.2.0
requests==2.31.0 